connector/requirements.txt
dbus-next==0.2.3
PyYAML>=6.0
